[FEAT] gif 다 만들어진 후 보이도록 수정
