+++ picpico-client/src/modules/step.mjs
@@ -1,45 +1,35 @@
-import {setDecoListInfo, setDoneDecoInfo, setMyDecoCanvasInfo, setRealResultInfo} from "../slice/decoInfo.js";
-import {setDecoInfo, setGalleryInfo, setPicBoothInfo, setSelectionInfo} from "../slice/picpicoInfo.js";
-import {setImgListInfo} from "../slice/selectionInfo.js";
+import { setDecoListInfo, setDoneDecoInfo, setMyDecoCanvasInfo, setRealResultInfo } from "../slice/decoInfo.js";
+import { setDecoInfo, setGalleryInfo, setPicBoothInfo, setSelectionInfo } from "../slice/picpicoInfo.js";
+import { setImgListInfo } from "../slice/selectionInfo.js";
 import store from "../store.js";
-import {socket} from "./sockets.mjs";
-import {ResultImage, Sticker} from "../modules/resultCanvas.mjs";
-import {gifTest} from "../test/resultTest.mjs";
+import { socket } from "./sockets.mjs";
+import { ResultImage, Sticker } from "../modules/resultCanvas.mjs";
+import { gifTest } from "../test/resultTest.mjs";
 
 export const onDoneTakeEvent = imgArr => {
     console.log("picBooth done~~", imgArr);
-    store.dispatch(setImgListInfo({value: imgArr}));
-    store.dispatch(setPicBoothInfo({value: false}));
-    store.dispatch(setSelectionInfo({value: true}));
+    store.dispatch(setImgListInfo({ value: imgArr }));
+    store.dispatch(setPicBoothInfo({ value: false }));
+    store.dispatch(setSelectionInfo({ value: true }));
 };
 
 export const onDonePickEvent = imgArr => {
-    store.dispatch(setDecoListInfo({value: imgArr}));
-    store.dispatch(setSelectionInfo({value: false}));
-    store.dispatch(setDecoInfo({value: true}));
-    store.dispatch(setMyDecoCanvasInfo({value: Object.keys(imgArr)[0]}));
+    store.dispatch(setDecoListInfo({ value: imgArr }));
+    store.dispatch(setSelectionInfo({ value: false }));
+    store.dispatch(setDecoInfo({ value: true }));
+    store.dispatch(setMyDecoCanvasInfo({ value: Object.keys(imgArr)[0] }));
 };
 
 export const onDoneDecoEvent = () => {
     const result = store.getState().decoInfo.resultList;
-    socket.emit("submit_deco", result);
+    socket.emit("submit_deco");
     // store.dispatch(setDecoInfo({ value: false }));
     // store.dispatch(setGalleryInfo({ value: true }));
 };
 
-export const onSubmitDecoEvent = async realResult => {
+export const onSubmitDecoEvent = () => {
     console.log("on submit deco");
-    store.dispatch(setRealResultInfo({value: realResult}));
 
-    store.dispatch(setDecoInfo({value: false}));
-    store.dispatch(setGalleryInfo({value: true}));
-    console.log("back to the first");
-
-    // for (let i = 0; i < realResult.length; i++) {
-    //   const curRealResult = realResult[i];
-
-    // resultImages.push(new ResultImage(curRealResult["resultUrl"], curRealResult["stickers"]));
-    // }
-
-    console.log("realResult:", realResult);
+    store.dispatch(setDecoInfo({ value: false }));
+    store.dispatch(setGalleryInfo({ value: true }));
 };
